requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
redis==5.0.3
pika==1.3.2
pyyaml==6.0.1
colorama==0.4.6